get validation pylint compliant
