+++ validator/delphi_validator/datafetcher.py
@@ -4,6 +4,7 @@ Functions to get CSV filenames and data.
 """
 
 import re
+import threading
 from os import listdir
 from os.path import isfile, join
 from itertools import product
@@ -11,10 +12,70 @@ import pandas as pd
 import numpy as np
 
 import covidcast
-from .errors import APIDataFetchError
+from .errors import APIDataFetchError, ValidationError
 
 filename_regex = re.compile(
     r'^(?P<date>\d{8})_(?P<geo_type>\w+?)_(?P<signal>\w+)\.csv$')
+
+
+def make_date_filter(start_date, end_date):
+    """
+    Create a function to return a boolean of whether a filename of appropriate
+    format contains a date within (inclusive) the specified date range.
+
+    Arguments:
+        - start_date: datetime date object
+        - end_date: datetime date object
+
+    Returns:
+        - Custom function object
+    """
+    # Convert dates from datetime format to int.
+    start_code = int(start_date.strftime("%Y%m%d"))
+    end_code = int(end_date.strftime("%Y%m%d"))
+
+    def custom_date_filter(match):
+        """
+        Return a boolean of whether a filename of appropriate format contains a date
+        within the specified date range.
+
+        Arguments:
+            - match: regex match object based on filename_regex applied to a filename str
+
+        Returns:
+            - boolean
+        """
+        # If regex match doesn't exist, current filename is not an appropriately
+        # formatted source data file.
+        if not match:
+            return False
+
+        # Convert date found in CSV name to int.
+        code = int(match.groupdict()['date'])
+
+        # Return boolean True if current file date "code" is within the defined date range.
+        return start_code <= code <= end_code
+
+    return custom_date_filter
+
+
+def load_all_files(export_dir, start_date, end_date):
+    """Load all files in a directory.
+    Parameters
+    ----------
+    export_dir: str
+        directory from which to load files
+
+    Returns
+    -------
+    loaded_data: List[Tuple(str, re.match, pd.DataFrame)]
+        triples of filenames, filename matches with the geo regex, and the data from the file
+    """
+    export_files = read_filenames(export_dir)
+    date_filter = make_date_filter(start_date, end_date)
+
+    # Make list of tuples of CSV names and regex match objects.
+    return [(f, m, load_csv(join(export_dir, f))) for (f, m) in export_files if date_filter(m)]
 
 
 def read_filenames(path):
@@ -95,3 +156,60 @@ def fetch_api_reference(data_source, start_date, end_date, geo_type, signal_type
     ).reindex(columns=column_names)
 
     return api_df
+
+
+def get_one_api_df(data_source, min_date, max_date,
+                    geo_type, signal_type,
+                    api_semaphore, dict_lock, output_dict):
+    """
+    Pull API data for a single geo type-signal combination. Raises
+    error if data couldn't be retrieved. Saves data to data dict.
+    """
+    api_semaphore.acquire()
+
+    # Pull reference data from API for all dates.
+    try:
+        geo_sig_api_df_or_error = fetch_api_reference(
+            data_source, min_date, max_date, geo_type, signal_type)
+
+    except APIDataFetchError as e:
+        geo_sig_api_df_or_error = ValidationError(
+                ("api_data_fetch_error", geo_type, signal_type), None, e)
+
+    api_semaphore.release()
+
+    # Use a lock so only one thread can access the dictionary.
+    dict_lock.acquire()
+    output_dict[(geo_type, signal_type)] = geo_sig_api_df_or_error
+    dict_lock.release()
+
+def threaded_api_calls(min_date, max_date, geo_signal_combos, n_threads=32):
+    """
+    Get data from API for all geo-signal combinations in a threaded way
+    to save time.
+    """
+    if n_threads > 32:
+        n_threads = 32
+        print("Warning: Don't run more than 32 threads at once due "
+                + "to API resource limitations")
+
+    output_dict = dict()
+    dict_lock = threading.Lock()
+    api_semaphore = threading.Semaphore(value=n_threads)
+
+    thread_objs = [threading.Thread(
+        target=get_one_api_df, args=(min_date, max_date,
+                                     geo_type, signal_type,
+                                     api_semaphore,
+                                     dict_lock, output_dict)
+    ) for geo_type, signal_type in geo_signal_combos]
+
+    # Start all threads.
+    for thread in thread_objs:
+        thread.start()
+
+    # Wait until all threads are finished.
+    for thread in thread_objs:
+        thread.join()
+
+    return output_dict
